feat(SPV-848): webhook features to a separate package
