+++ examples/webhooks/webhooks.go
@@ -11,13 +11,21 @@ import (
 
 	walletclient "github.com/bitcoin-sv/spv-wallet-go-client"
 	"github.com/bitcoin-sv/spv-wallet-go-client/examples"
+	"github.com/bitcoin-sv/spv-wallet-go-client/notifications"
 )
 
 func main() {
 	defer examples.HandlePanic()
 
-	client := walletclient.NewWithAdminKey("http://localhost:3003/v1", "xprv9s21ZrQH143K2pmNeAHBzU4JHNDaFaPTbzKbBCw55ErhMDLsxDwKqcaDVV3PwmEmRZa9qUaU261iJaUx8eBiBF77zrPxTH8JGXC7LZQnsgA")
-	wh := walletclient.NewWebhook(context.Background(), client, "http://localhost:5005/notification", "Authorization", "this-is-the-token", 3)
+	client := walletclient.NewWithAdminKey("http://localhost:3003/v1", examples.ExampleAdminKey)
+	//"Authorization", "this-is-the-token", 3
+	wh := notifications.NewWebhook(
+		context.Background(),
+		client,
+		"http://localhost:5005/notification",
+		notifications.WithToken("Authorization", "this-is-the-token"),
+		notifications.WithProcessors(3),
+	)
 	err := wh.Subscribe(context.Background())
 	if err != nil {
 		panic(err)
@@ -25,14 +33,14 @@ func main() {
 
 	http.Handle("/notification", wh.HTTPHandler())
 
-	if err := walletclient.RegisterHandler(wh, func(gpe *walletclient.NumericEvent) {
+	if err = notifications.RegisterHandler(wh, func(gpe *notifications.NumericEvent) {
 		time.Sleep(50 * time.Millisecond) // simulate processing time
 		fmt.Printf("Processing event-numeric: %d\n", gpe.Numeric)
 	}); err != nil {
 		panic(err)
 	}
 
-	if err := walletclient.RegisterHandler(wh, func(gpe *walletclient.StringEvent) {
+	if err = notifications.RegisterHandler(wh, func(gpe *notifications.StringEvent) {
 		time.Sleep(50 * time.Millisecond) // simulate processing time
 		fmt.Printf("Processing event-string: %s\n", gpe.Value)
 	}); err != nil {
